src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/gesture-tapHighlight-shadow-tree-expected.txt

CONSOLE ERROR: line 17: Uncaught TypeError: document.querySelector(...).createShadowRoot is not a function
CONSOLE ERROR: line 22: Uncaught TypeError: Cannot read property 'getBoundingClientRect' of undefined
